Fix: Update top_predictions type hint in PredictionResponse to use List[float] instead of List[int]

+++ bst_export_serverless/serverless_api.py
@@ -86,7 +86,7 @@ class PredictionResponse(BaseModel):
     inference_time: float
     predictions: List[List[float]]
     probabilities: List[List[float]]
-    top_predictions: Dict[str, List[List[int]]]
+    top_predictions: Dict[str, List[List[float]]]
     metadata: Dict[str, Any]
 
 class ErrorResponse(BaseModel):
